Midterm
I want to work with some gifs of early 90’s era music videos and more
contemporary audio. I want to combine the older video footage with
newer songs and have the visual and audio choose from an array and
match up randomly on click. I’ve created a number of gifs and placed
them in an array.